+++ perf/tests/test_perf_cli.py
@@ -8,6 +8,7 @@ from perf.tests import unittest
 
 
 TELCO = os.path.join(os.path.dirname(__file__), 'telco.json')
+TRACK_MEMORY = os.path.join(os.path.dirname(__file__), 'track_memory.json')
 
 
 class BaseTestCase(object):
@@ -313,6 +314,20 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
         stdout = self.run_command('dump', TELCO)
         self.assertIn(textwrap.dedent(expected).strip(), stdout)
 
+    def test_dump_track_memory(self):
+        expected = """
+            Run 1: calibrate the number of loops: 2^15
+            - calibrate 1: 7188.0 kB (loops: 2^15)
+            Run 2: 0 warmups, 1 value, 2^15 loops
+            - value 1: 7188.0 kB
+            Run 3: 0 warmups, 1 value, 2^15 loops
+            - value 1: 7192.0 kB
+            Run 4: 0 warmups, 1 value, 2^15 loops
+            - value 1: 7208.0 kB
+        """
+        stdout = self.run_command('dump', TRACK_MEMORY)
+        self.assertIn(textwrap.dedent(expected).strip(), stdout)
+
     def test_dump_quiet(self):
         expected = """
             Run 2:
@@ -523,6 +538,33 @@ class TestConvert(BaseTestCase, unittest.TestCase):
         self.assertEqual(bench2.get_values(), (4.0,))
         self.assertEqual(bench3.get_values(), (1.0, 2.0, 3.0, 5.0))
         self.assertEqual(bench4.get_values(), (1.0, 3.0, 5.0))
+
+    def _check_track_memory(self, track_option):
+        with tests.temporary_file() as tmp_name:
+            self.run_command('timeit',
+                             track_option,
+                             '-p3', '-w0', '-l5', '-n3',
+                             '[1,2]*1000',
+                             '-o', tmp_name)
+            bench = perf.Benchmark.load(tmp_name)
+
+        self.assertEqual(bench.get_nrun(), 3)
+        for run in bench.get_runs():
+            self.assertEqual(run.warmups, ())
+            self.assertEqual(len(run.values), 1)
+            self.assertIsInstance(run.values[0], int)
+            self.assertEqual(run.get_loops(), 5)
+
+    def test_track_memory(self):
+        self._check_track_memory('--track-memory')
+
+    def test_tracemalloc(self):
+        try:
+            import tracemalloc   # noqa
+        except ImportError:
+            self.skipTest('tracemalloc module not available')
+
+        self._check_track_memory('--tracemalloc')
 
 
 if __name__ == "__main__":